Rename Templates to Ramhorns, moved to top level
