+++ ramhorns/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "ramhorns"
 version = "0.8.1"
-authors = ["Maciej Hirsz <maciej.hirsz@pm.me>"]
+authors = ["Maciej Hirsz <hello@maciej.codes>"]
 license = "GPL-3.0"
 edition = "2018"
 description = "Experimental Mustache-like templating engine."
@@ -12,7 +12,7 @@ keywords = ["html", "template", "mustache"]
 categories = ["template-engine"]
 
 [dependencies]
-beef = "0.2.1"
+beef = "0.3.0"
 fnv = "1.0"
 pulldown-cmark = "0.7"
 ramhorns-derive = { version = "0.8.0", optional = true }
