+++ ramhorns/src/lib.rs
@@ -76,6 +76,11 @@
 
 #![warn(missing_docs)]
 
+use std::path::{Path, PathBuf};
+
+use beef::Cow;
+use rustc_hash::FxHashMap;
+
 mod cmark;
 mod content;
 mod error;
@@ -86,7 +91,108 @@ pub mod encoding;
 
 pub use content::Content;
 pub use error::Error;
-pub use template::{Section, Template, Templates};
+pub use template::{Section, Template};
 
 #[cfg(feature = "export_derive")]
 pub use ramhorns_derive::Content;
+
+/// A safe wrapper around a `HashMap` containing preprocessed templates
+/// of the type `Template`, accesible by their name.
+pub struct Ramhorns {
+    partials: FxHashMap<Cow<'static, str>, Template<'static>>,
+    dir: PathBuf,
+}
+
+impl Ramhorns {
+    /// Loads all the `.html` files as templates from the given folder into a hashmap, making them
+    /// accessible via their path, joining partials as required
+    /// ```no_run
+    /// # use ramhorns::Ramhorns;
+    /// let tpls = Ramhorns::from_folder("./templates").unwrap();
+    /// let content = "I am the content";
+    /// let rendered = tpls.get("hello.html").unwrap().render(&content);
+    /// ```
+    pub fn from_folder<P: AsRef<Path>>(dir: P) -> Result<Self, Error> {
+        let mut templates = Ramhorns::lazy(dir)?;
+
+        fn load_folder(path: &Path, templates: &mut Ramhorns) -> Result<(), Error> {
+            for entry in std::fs::read_dir(path)? {
+                let path = entry?.path();
+                if path.is_dir() {
+                    load_folder(&path, templates)?;
+                } else if path.extension().unwrap_or_else(|| "".as_ref()) == "html" {
+                    let name = path
+                        .strip_prefix(&templates.dir)
+                        .unwrap_or(&path)
+                        .to_string_lossy();
+
+                    if !templates.partials.contains_key(&*name) {
+                        let template = Template::load(std::fs::read_to_string(&path)?, templates)?;
+                        templates
+                            .partials
+                            .insert(name.into_owned().into(), template);
+                    }
+                }
+            }
+            Ok(())
+        }
+        load_folder(&templates.dir.clone(), &mut templates)?;
+
+        Ok(templates)
+    }
+
+    /// Create a new empty mapping of files to template for a given folder.
+    /// ```no_run
+    /// # use ramhorns::Ramhorns;
+    /// let mut tpls = Ramhorns::lazy("./templates").unwrap();
+    /// let content = "I am the content";
+    /// let rendered = tpls.from_file("hello.html").unwrap().render(&content);
+    /// ```
+    pub fn lazy<P: AsRef<Path>>(dir: P) -> Result<Self, Error> {
+        Ok(Ramhorns {
+            partials: FxHashMap::default(),
+            dir: dir.as_ref().canonicalize()?,
+        })
+    }
+
+    /// Get the template with the given name, if it exists.
+    pub fn get<S>(&self, name: &S) -> Option<&Template<'static>>
+    where
+        for<'a> Cow<'a, str>: std::borrow::Borrow<S>,
+        S: std::hash::Hash + AsRef<Path> + Eq + ?Sized,
+    {
+        self.partials.get(name)
+    }
+
+    /// Get the template with the given name. If the template doesn't exist,
+    /// it will be loaded from file and parsed first.
+    ///
+    /// Use this method in tandem with `Ramhorns::lazy`.
+    pub fn from_file(&mut self, name: &str) -> Result<&Template<'static>, Error> {
+        self.load_internal(name.to_owned().into())
+    }
+
+    fn load_internal(&mut self, name: Cow<'static, str>) -> Result<&Template<'static>, Error> {
+        let n: &str = &name;
+        if !self.partials.contains_key(n) {
+            let path = self.dir.join(n).canonicalize()?;
+
+            if !path.starts_with(&self.dir) {
+                return Err(Error::IllegalPartial(n.into()));
+            }
+            let template = Template::load(std::fs::read_to_string(&path)?, self)?;
+            self.partials.insert(name.clone(), template);
+        };
+        Ok(&self.partials[n])
+    }
+}
+
+pub(crate) trait Partials<'tpl> {
+    fn get_partial(&mut self, name: &'tpl str) -> Result<&Template<'tpl>, Error>;
+}
+
+impl Partials<'static> for Ramhorns {
+    fn get_partial(&mut self, name: &'static str) -> Result<&Template<'static>, Error> {
+        self.load_internal(Cow::borrowed(name))
+    }
+}
